refactor: Split voting hooks
