+++ src/views/Voting/components/Proposals/Proposals.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'contexts/Localization'
 import Container from 'components/Layout/Container'
 import { useAppDispatch } from 'state'
 import { fetchProposals } from 'state/voting'
-import { useGetProposalLoadingStatus, useGetProposals } from 'state/hooks'
+import { useGetProposalLoadingStatus, useGetProposals } from 'state/voting/hooks'
 import { ProposalState, ProposalType, VotingStateLoadingStatus } from 'state/types'
 import { filterProposalsByState, filterProposalsByType } from '../../helpers'
 import BreadcrumbLink from '../BreadcrumbLink'
